Outcomment allied vision import and interface test for the manta class.
The import created problems with CI, because the VIMBA driver is not
installed on the server.
Since the pymba package is not message based (no write or read methods
needed by a user) this device does not count as a message based device
and therefore doesnt need to fullfill the device-interface conditions.