Changed filename of test module for allied vision device
This is done because in this way the unittest discover command used by travisCI
will not find the test. The test can still be used manually, but
travisCI can only use it if the Vimba driver from Allied Vision is installed
and I currently don't know how to implement this for the travis server.